Fix some getModel() / parameter passing related test failures.
This is a continuation of the work to fix issue feathersjs-ecosystem#234.
Some service calls were still failing to pass through params.sequelize,
which again could cause issues for users that are extending
the Service class with their own getModel() method that
depends upon certain parameters being passed through to it.