Tighten up some options.Model/Sequelize/getModel related code.
Firstly, some background:
feathers-sequelize allows for overriding of its getModel method
in order to let users implement their own model provider schemes;

Overriding getModel() in this way allows for a different Model to be programatically
provided any time getModel() is called.
This can allow for multi-tenancy setups where different customer models/databases
are used depending upon different criteria, such as what hostname they're connecting to,
where they are connecting from, or whatever the programmer desires.

Currently, the feathers-sequelize Service class requires that a single Model be provided
in the options passed to its constructor. The problem is, when using a getModel() override
as described above, a single Model is probably not known at the time that the feathers-sequelize
Service constructor is called.

Therefore, a Model probably shouldn't be required at construction time.

This changeset makes providing a Model optional, adds an alternative option attribute called 'Sequelize'
for providing the Sequelize class (which is currently used to derive the default Ops), and
adds some warning messages inside of the Model getter to dissaude usage of it altogether other than
for coverage tests, since using this getter doesn't allow for proper getModel() overriding.