Fix issue with patch when using pagination by default
Currently patch does not disable pagination when gathering
a list of IDs to be modified.

If pagination is enabled by default for a service, this
causes the request to fail when patching multiple records
with a query.

This commit updates the patch function to use getOrFind
rather than find, with the former explicitly disabling
pagination.